Move palette detection into SLPEncoder
